adjusted model attributes

+++ Models/Books.cs
@@ -9,14 +9,14 @@ namespace Mission_09_alley725.Models
 {
     public partial class Book
     {
-        public long BookId { get; set; }
+        public int BookId { get; set; }
         public string Title { get; set; }
         public string Author { get; set; }
         public string Publisher { get; set; }
         public string Isbn { get; set; }
         public string Classification { get; set; }
         public string Category { get; set; }
-        public long PageCount { get; set; }
+        public int PageCount { get; set; }
         public double Price { get; set; }
     }
 }
